Move PageFooter to separate file
